requirements.txt
aniso8601==1.3.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-RESTful==0.3.6
gevent==1.2.2
greenlet==0.4.12
idna==2.6
itsdangerous==0.24
Jinja2==2.10
locustio==0.8.1
MarkupSafe==1.0
msgpack-python==0.4.8
python-dateutil==2.6.1
pytz==2017.3
pyzmq==16.0.3
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
psycopg2==2.7.3.2
SQLAlchemy==1.1.15
Flask-SQLAlchemy==2.3.2
sqlalchemy-utils==0.32.21
redis==2.10.6
Flask-Redis==0.3.0
gunicorn==19.7.1
mock==2.0.0
